third_party/blink/web_tests/fast/mediastream/constructors-expected.txt

Tests the JSEP PeerConnection related constructors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof RTCPeerConnection === 'function' is true
PASS typeof RTCSessionDescription === 'function' is true
PASS typeof RTCIceCandidate === 'function' is true
PASS RTCPeerConnection() threw exception TypeError: Failed to construct 'RTCPeerConnection': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS RTCSessionDescription() threw exception TypeError: Failed to construct 'RTCSessionDescription': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS RTCIceCandidate() threw exception TypeError: Failed to construct 'RTCIceCandidate': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS new RTCPeerConnection({iceServers:[{url:'stun:foobar.com:12345'}]}, null); did not throw exception.
PASS new RTCSessionDescription({type:'offer',sdp:'foobar'}); did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
